requirements.txt
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
sentry-sdk==2.14.0
urllib3==2.2.3
Werkzeug==3.0.4

